Revert "refactor: `Radio` component styling (#488)"
This reverts commit 181c3979a70c560b3bd05a4281c70467544c17e5.
