--- packages/native-ui/src/components/Icons/index.ts
@@ -1,1 +1,0 @@
-export { default as CircleIcon } from './CircleIcon';
