--- packages/native-ui/src/components/Radio/index.ts
@@ -1,7 +1,0 @@
-export {
-  default as Radio,
-  RadioGroup as RadioGroup,
-  RadioIndicator as RadioIndicator,
-  RadioLabel as RadioLabel,
-  RadioIcon as RadioIcon,
-} from './Radio';
