+++ packages/native-ui/src/components/RadioIndicator/index.ts
@@ -1,0 +1,1 @@
+export { default } from './RadioIndicator';
