+++ packages/native-ui/src/components/index.ts
@@ -2,13 +2,13 @@
 export * from './IconButton';
 export * from './Alert';
 export { default as AnimatedOutline } from './AnimatedOutline';
+export { default as RadioIndicator } from './RadioIndicator';
 export * from './Button';
 export * from './Spinner';
 export * from './List';
 export * from './Skeleton';
 export * from './Badge';
 export * from './Checkbox';
-export * from './Radio';
 
 // Gluestack UI
 export {
@@ -39,6 +39,7 @@ export { Checkbox, CheckboxGroup, CheckboxLabel } from '@gluestack-ui/themed';
 
 export { Link, LinkText } from '@gluestack-ui/themed';
 export { Pressable } from '@gluestack-ui/themed';
+export { Radio, RadioGroup, RadioLabel, RadioIcon } from '@gluestack-ui/themed';
 
 export { Image } from '@gluestack-ui/themed';
