--- packages/native-ui/src/types/states.ts
@@ -1,11 +1,0 @@
-export interface States {
-  readonly active?: boolean;
-  readonly checked?: boolean;
-  readonly disabled?: boolean;
-  readonly focus?: boolean;
-  readonly focusVisible?: boolean;
-  readonly hover?: boolean;
-  readonly indeterminate?: boolean;
-  readonly invalid?: boolean;
-  readonly readonly?: boolean;
-}
